event-timing/event-click-counts.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Event Timing: eventCounts.</title>
<div id='div'>Click me</div>
<button id='button'>Click me</div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=resources/event-timing-test-utils.js></script>
<script>
  promise_test( t => {
    assert_precondition(window.EventCounts, "Event Counts isn't supported");
    function testClicks(expectedCount, resolve) {
      const clickCount = performance.eventCounts.get('click');
      if (!clickCount || clickCount < expectedCount) {
        t.step_timeout(function() {
          testClicks(expectedCount, resolve);
        }, 5);
        return;
      }
      assert_equals(clickCount, expectedCount,'Incorrect click count.');
      assert_equals(performance.eventCounts.get('mousedown'), expectedCount, 'Incorrect mousedown count');
      assert_equals(performance.eventCounts.get('mouseup'), expectedCount, 'Incorrect mouseup count.');
      assert_equals(performance.eventCounts.get('mouseover'), expectedCount, 'Incorrect mouseover count.');
      // There should be no other input type recorded.
      assert_equals(performance.eventCounts.size, 4, 'There should only be 4 types observed.');
      resolve();
    }
    function promiseClicks(expectedCount) {
      return new Promise(resolve => {
        testClicks(1, resolve)
      });
    }

    assert_equals(performance.eventCounts.size, 0);
    test_driver.click(document.getElementById('div'));
    return promiseClicks(1).then(() => {
      test_driver.click(document.getElementById('button'));
      return promiseClicks(2);
    }).then(() => {
      test_driver.click(document.getElementById('div'));
      return promiseClicks(3);
    });
  })
</script>
</html>
